Moved models to external folder
