+++ src/main/kotlin/org/ionproject/integration/job/ISELTimetable.kt
@@ -1,7 +1,7 @@
 package org.ionproject.integration.job
 
 import org.ionproject.integration.format.implementations.ISELTimetableFormatChecker
-import org.ionproject.integration.model.internal.timetable.TimetableTeachers
+import org.ionproject.integration.model.external.timetable.TimetableTeachers
 import org.ionproject.integration.model.internal.timetable.UploadType
 import org.ionproject.integration.model.internal.timetable.isel.RawData
 import org.ionproject.integration.step.chunkbased.FormatVerifierStepBuilder
