+++ src/main/kotlin/org/ionproject/integration/model/external/timetable/Label.kt
@@ -1,0 +1,5 @@
+package org.ionproject.integration.model.external.timetable
+
+data class Label(
+    val acr: String
+)
