+++ src/main/kotlin/org/ionproject/integration/model/external/timetable/Programme.kt
@@ -1,0 +1,5 @@
+package org.ionproject.integration.model.external.timetable
+
+data class Programme(
+    var name: String = ""
+)
