+++ src/main/kotlin/org/ionproject/integration/model/external/timetable/Teacher.kt
@@ -1,0 +1,5 @@
+package org.ionproject.integration.model.external.timetable
+
+data class Teacher(
+    val name: String
+)
