+++ src/main/kotlin/org/ionproject/integration/model/external/timetable/TimetableTeachers.kt
@@ -1,0 +1,9 @@
+package org.ionproject.integration.model.external.timetable
+
+import org.ionproject.integration.model.external.timetable.CourseTeacher
+import org.ionproject.integration.model.external.timetable.Timetable
+
+data class TimetableTeachers(
+    var timetable: List<Timetable> = listOf(),
+    var teachers: List<CourseTeacher> = listOf()
+)
