--- src/main/kotlin/org/ionproject/integration/model/internal/timetable/Label.kt
@@ -1,5 +1,0 @@
-package org.ionproject.integration.model.internal.timetable
-
-data class Label(
-    val acr: String
-)
