--- src/main/kotlin/org/ionproject/integration/model/internal/timetable/Programme.kt
@@ -1,5 +1,0 @@
-package org.ionproject.integration.model.internal.timetable
-
-data class Programme(
-    var name: String = ""
-)
