--- src/main/kotlin/org/ionproject/integration/model/internal/timetable/Teacher.kt
@@ -1,5 +1,0 @@
-package org.ionproject.integration.model.internal.timetable
-
-data class Teacher(
-    val name: String
-)
