--- src/main/kotlin/org/ionproject/integration/model/internal/timetable/Timetable.kt
@@ -1,10 +1,0 @@
-package org.ionproject.integration.model.internal.timetable
-
-data class Timetable(
-    var school: School = School(),
-    var programme: Programme = Programme(),
-    var calendarTerm: String = "",
-    var calendarSection: String = "",
-    var language: String = "",
-    var courses: List<Course> = listOf()
-)
