--- src/main/kotlin/org/ionproject/integration/model/internal/timetable/TimetableTeachers.kt
@@ -1,6 +1,0 @@
-package org.ionproject.integration.model.internal.timetable
-
-data class TimetableTeachers(
-    var timetable: List<Timetable> = listOf(),
-    var teachers: List<CourseTeacher> = listOf()
-)
