+++ src/main/kotlin/org/ionproject/integration/service/implementations/CoreService.kt
@@ -7,8 +7,8 @@ import java.net.http.HttpResponse
 import java.time.Duration
 import org.ionproject.integration.config.AppProperties
 import org.ionproject.integration.model.internal.core.CoreResult
-import org.ionproject.integration.model.internal.timetable.CourseTeacher
-import org.ionproject.integration.model.internal.timetable.Timetable
+import org.ionproject.integration.model.external.timetable.CourseTeacher
+import org.ionproject.integration.model.external.timetable.Timetable
 import org.ionproject.integration.service.interfaces.ICoreService
 import org.ionproject.integration.utils.HttpUtils
 import org.ionproject.integration.utils.JsonUtils
