+++ src/test/kotlin/org/ionproject/integration/service/implementations/CoreServiceTests.kt
@@ -6,9 +6,9 @@ import java.net.http.HttpResponse
 import java.time.Duration
 import org.ionproject.integration.config.AppProperties
 import org.ionproject.integration.model.internal.core.CoreResult
-import org.ionproject.integration.model.internal.timetable.CourseTeacher
-import org.ionproject.integration.model.internal.timetable.School
-import org.ionproject.integration.model.internal.timetable.Timetable
+import org.ionproject.integration.model.external.timetable.CourseTeacher
+import org.ionproject.integration.model.external.timetable.School
+import org.ionproject.integration.model.external.timetable.Timetable
 import org.ionproject.integration.utils.HttpUtils
 import org.ionproject.integration.utils.orThrow
 import org.junit.jupiter.api.Assertions.assertEquals
